# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is Product Promo System for US Market .",
        "title": "Product Promo System API - US Market",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/product/{id}": {
            "get": {
                "description": "Get method for particular product information from server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get particular Product's information from the server",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "description": "Get method for listing all the products from server",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "List all Products information from the server",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}